refactor: split logger into modules
